Factorize LineEdge/TwtEdge common code
