+++ single-line-diagram-core/src/main/java/com/powsybl/sld/model/LineEdge.java
@@ -9,8 +9,6 @@ package com.powsybl.sld.model;
 import com.fasterxml.jackson.core.JsonGenerator;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 /**
@@ -19,7 +17,6 @@ import java.util.Objects;
 public class LineEdge extends AbstractBranchEdge {
 
     private final String lineId;
-    private List<Point> points = new ArrayList<>();
 
     public LineEdge(String lineId, Node node1, Node node2) {
         super(node1, node2);
@@ -31,45 +28,11 @@ public class LineEdge extends AbstractBranchEdge {
     }
 
     public void addPoint(double x, double y) {
-        points.add(new Point(x, y));
-    }
-
-    public void setPoints(List<Point> points) {
-        this.points = Objects.requireNonNull(points);
-    }
-
-    public List<Point> getPoints() {
-        return points;
+        snakeLine.add(new Point(x, y));
     }
 
     @Override
-    public List<Point> getSnakeLine() {
-        return getPoints();
-    }
-
-    @Override
-    public void setSnakeLine(List<Point> snakeLine) {
-        points = Objects.requireNonNull(snakeLine);
-    }
-
-    @Override
-    void writeJson(JsonGenerator generator) throws IOException {
-        writeJson(generator, false);
-    }
-
-    void writeJson(JsonGenerator generator, boolean generateCoordsInJson) throws IOException {
-        generator.writeStartObject();
+    public void writeIdJson(JsonGenerator generator) throws IOException {
         generator.writeStringField("lineId", lineId);
-        generator.writeArrayFieldStart("nodes");
-        super.writeJson(generator);
-        generator.writeEndArray();
-        if (generateCoordsInJson) {
-            generator.writeArrayFieldStart("points");
-            for (Point point : points) {
-                point.writeJson(generator);
-            }
-            generator.writeEndArray();
-        }
-        generator.writeEndObject();
     }
 }
